+++ nylas/client/restful_models.py
@@ -150,6 +150,7 @@ class Message(NylasAPIObject):
         "_folder",
         "_labels",
         "headers",
+        "reply_to",
     ]
     datetime_attrs = {"received_at": "date"}
     datetime_filter_attrs = {